# Conflicts:
#	kmaooad-functions/src/main/java/edu/kmaooad/handler/impl/group/AddGroupDialog.java
#	kmaooad-functions/src/main/java/edu/kmaooad/handler/impl/group/DeleteGroupDialog.java
#	kmaooad-functions/src/main/java/edu/kmaooad/handler/impl/group/UpdateGroupDialog.java
#	kmaooad-functions/src/main/java/edu/kmaooad/handler/impl/group/assign/AssignActivitiesToGroupHandler.java
#	kmaooad-functions/src/main/java/edu/kmaooad/model/Activity.java
#	kmaooad-functions/src/main/java/edu/kmaooad/service/ActivityService.java

+++ kmaooad-functions/src/main/java/edu/kmaooad/handler/impl/group/assign/AssignActivitiesToGroupHandler.java
@@ -11,6 +11,10 @@ import edu.kmaooad.model.HandlerResponse;
 import edu.kmaooad.model.UserRequest;
 import edu.kmaooad.service.ActivityService;
 import edu.kmaooad.service.GroupService;
+import edu.kmaooad.service.ServiceException;
+import edu.kmaooad.service.TelegramService;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
 import edu.kmaooad.service.TelegramService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -43,20 +47,27 @@ public class AssignActivitiesToGroupHandler implements UserRequestHandler {
         return UserRequestHandler.isTextMessage(userRequest.getUpdate()) &&
                 userRequest.getUserSession().getConversationState().equals(ConversationState.WAITING_FOR_GROUP_ACTIVITY_ASSIGN_CHOICE);
     }
-    //todo
-    // take off comment
+
     @Override
     public HandlerResponse handle(UserRequest userRequest) {
         Group group = (Group) userRequest.getUserSession().getData().get(GroupConstants.GROUP_MAP_KEY);
         String text = userRequest.getUpdate().getMessage().getText();
-       /* if(text.contains(ASSIGNED)) {
-            String activityName = text.replace(ASSIGNED, "").trim();
-            Activity activity = activityService.getActivityByName(activityName);
-            groupService.deleteActivityGroup(activity.getId(), group.getId());
-        } else {
-            Activity activity = activityService.getActivityByName(text);
-            groupService.addActivityGroup(activity.getId(), group.getId());
-        }*/
+
+        try {
+            if (text.contains(ASSIGNED)) {
+                String activityName = text.replace(ASSIGNED, "").trim();
+                Activity activity = activityService.getActivityByName(activityName);
+                groupService.deleteActivityGroup(activity.getId(), group.getId());
+                log.info("Deleted activity " + activityName + " from group " + group.getName());
+            } else {
+                Activity activity = activityService.getActivityByName(text);
+                groupService.addActivityGroup(activity, group.getId());
+                log.info("Added activity " + activity.getName() + " to group " + group.getName());
+            }
+        } catch (ServiceException e) {
+            e.printStackTrace();
+            log.error("Failed to delete or add activity " + text + " to group" + group.getName());
+        }
 
         userRequest.getUserSession().getData().put(GROUP_MAP_KEY, groupService.getGroupById(group.getId()));
         userRequest.getUserSession().setConversationState(ConversationState.WAITING_FOR_GROUP_TO_ASSIGN_CHOICE);
